% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_MRF.R
\name{bootstrap_MRF}
\alias{bootstrap_MRF}
\title{Bootstrap observations to estimate MRF parameter coefficients}
\usage{
bootstrap_MRF(data, n_bootstraps, sample_seed, min_lambda1, max_lambda1,
  by_lambda1, lambda2, separate_min, n_nodes, n_cores, n_covariates)
}
\arguments{
\item{data}{Dataframe. The input data where the \code{n_nodes}
left-most variables are binary occurrences to be represented by nodes in the graph.
Note that \code{NA}'s are allowed for covariates. If present, these missing values
will be imputed from the distribution \code{rnorm(mean = 0, sd = 1)}, which assumes that
all covariates are scaled and centred (i.e. by using the function
\code{\link[base]{scale}} or similar)}

\item{n_bootstraps}{Positive integer. Represents the total number of bootstrap samples
to test at each lambda1 value. Default is 100.}

\item{sample_seed}{Numeric. Used as the seed value for generating bootstrap replicates, allowing
users to generate replicated datasets on different systems. Default is a random seed}

\item{min_lambda1}{Positive numeric. The lowest l1−regularization value to be tested}

\item{max_lambda1}{Positive numeric. The highest l1−regularization to be tested}

\item{by_lambda1}{Positive numeric. The increment of the l1 test sequence. The test sequence is generated by calling
\code{lambda1_seq = seq(min_lambda1, max_lambda1, by_lambda1)}}

\item{lambda2}{Numeric (>= 0). Value for l2−regularization, where larger values lead
to stronger shrinking of coefficient magnitudes. Default is 0, but larger values
may be necessary for large or particularly sparse datasets}

\item{separate_min}{Logical. If \strong{TRUE}, interaction coefficients will use the minimum absolute value of
the corresponding parameter estimates, which are taken from separate logistic regressions,
in the symmetric postprocessed coefficient matrix. Else use the maximum. Default is \strong{FALSE}}

\item{n_nodes}{Positive integer. The index of the last column in \code{data}
which is represented by a node in the final graph. Columns with index
greater than n_nodes are taken as covariates. Default is the number of
columns in \code{data}, corresponding to no additional covariates}

\item{n_cores}{Integer. The number of cores to spread the job across using
\code{\link[parallel]{makePSOCKcluster}}. Default is 1 (no parallelisation)}

\item{n_covariates}{Positive integer. The number of covariates in \code{data},
before cross-multiplication. Default is \code{ncol(data) - n_nodes}}
}
\value{
A \code{list} containing five objects:
\itemize{
  \item \code{lambda_results}: \code{list} of length \code{lambda1_seq} containing:
  \itemize{
  \item \code{key_covariates}: dataframes of important covariates (i.e. those retained in
  at least 90 percent of bootstrap replicates) at each l1 value
  \item \code{raw_coefs}: the raw estimated coefficients at each l1 value
  \item \code{indirect_coefs}: estimated higher order interaction coefficients at each l1 value
  }
  \item \code{direct_coef_means}: \code{dataframe} containing mean coefficient values taken from all
  bootstrapped models across the range of l1 values
  \item \code{direct_coef_upper90} and \code{direct_coef_lower90}: \code{dataframe}s
  containing coefficient 95 percent and 5 percent quantiles taken from all
  bootstrapped models across the range of l1 values
  \item \code{indirect_coef_mean}: \code{list} of matrices containing mean higher order coefficient values
  taken from all bootstrapped models across the range of l1 values
  \item \code{mean_key_coefs}: \code{list} of matrices of length \code{n_nodes}
  containing mean covariate coefficient values and their relative importances
  (using the formula \code{x^2 / sum (x^2)}
  taken from all bootstrapped models across the range of l1 values. Only coefficients
  with mean relative importances \code{>0.01} are returned. Note, relative importance should only
  be considered useful if all covariates are on a similar scale!
}
}
\description{
This function runs \code{\link{MRFcov}} models across a user-specified
range of l1 regularization values. To capture uncertainty in paramter esimates,
the dataset is bootstrapped a user-specified number of times at each l1 value.
}
\details{
\code{MRFcov} models are run across the specified sequence of \code{lambda1} values.
For each model, the \code{data} is bootstrapped by shuffling row observations, using
\code{dplyr::sample_n(data, nrow(data), TRUE)}, to account for uncertainty.
Parameter estimates from the set of bootstrapped models are summarised
to present confidence intervals.
}
\examples{
\dontrun{
data("Bird.parasites")
bootedCRF <- bootstrap_MRF(data = Bird.parasites,
                          n_nodes = 4, min_lambda1 = 0.5,
                          max_lambda1 = 1.25,
                          by_lambda1 = 0.25)}
}
\seealso{
\code{\link{MRFcov}},
\code{\link[penalized]{penalized}}
}
