% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.rosalia.MRF.R
\name{comp.rosalia.MRF}
\alias{comp.rosalia.MRF}
\title{Compare MRFcov and rosalia outputs}
\usage{
comp.rosalia.MRF(MRF, rosalia)
}
\arguments{
\item{MRF}{A fitted \code{\link{MRFcov}} model}

\item{rosalia}{A fitted \code{\link[rosalia]{rosalia}} model}
}
\value{
A \code{list} of two objects:
\itemize{
   \item \code{Positive_predictive_value}: The positive predictive value
   \item \code{Negative_predictive_value}: The negative predictive value
   }
}
\description{
This function compares coefficients estimated from binary
\code{\link[rosalia]{rosalia}} MRF models (fit by Maximum Likelihood) and
\code{\link{MRFcov}} models (fit by separate penalized regressions). Here,
\code{\link[rosalia]{rosalia}} coefficients are considered the \emph{true} values,
and \code{\link{MRFcov}} coefficients are then compared to assess positive and negative
predictive capacity of the separate penalized regression approach.
}
\seealso{
\code{\link{MRFcov}}, \code{\link[rosalia]{rosalia}}
}
\author{
Nicholas Clark: \url{nicholas.j.clark1214@gmail.com}
}
