% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countzero.R
\name{countzero}
\alias{countzero}
\title{Count proportions of non-zero coefficients}
\usage{
countzero(data, x, y)
}
\arguments{
\item{data}{Matrix of coefficients returned from a fitted \code{MRFcov} model}

\item{x}{Integer for indexing rows in \code{data}}

\item{y}{Integer for indexing columns in \code{data}}
}
\value{
Matrix containing proportions of replicates in which each coefficient was retained
}
\description{
This function counts the proportion of bootstrap estimates
in which each tested model coefficient is nonzero. Note that
this function is not called directly, but is used in the function
\code{\link{bootstrap_MRF}}
}
\seealso{
\code{\link{bootstrap_MRF}}
}
\author{
Nicholas Clark: \url{nicholas.j.clark1214@gmail.com}
}
