% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMRF_adjacency.R
\name{getMRF_adjacency}
\alias{getMRF_adjacency}
\title{Extract predicted adjacency matrices for a given dataset}
\usage{
getMRF_adjacency(data, MRF_mod)
}
\arguments{
\item{data}{Dataframe. The sample data where the
left-most variables are variables that are represented by nodes in the graph}

\item{MRF_mod}{A fitted \code{MRFcov} object}
}
\value{
A \code{list} of weighted, undirected \code{igraph} adjacency matrices of length
\code{n_predvals} (one matrix for each fake value of the covariate). The names of the returned
\code{list} represent the covariate value used for prediction
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}}
or \code{\link{bootstrap_MRF}} along with a
supplied sample dataset to generate linear predictions of species' occurrences / abundances.
The predections are then used to create a weighted, undirected \code{igraph} adjacency matrix.
}
\details{
Interaction parameters from \code{MRF_mod} are
extracted and converted into a predicted weighted, undirected adjacency matrix
using \code{\link[igraph]{graph.adjacency}}.
}
\examples{
data("Bird.parasites")
CRFmod <- MRFcov(data = Bird.parasites, n_nodes = 4, family = "binomial")
getMRF_adjacency(data = Bird.parasites[1:300, ], MRF_mod = CRFmod)


}
\seealso{
\code{\link{MRFcov}}, \code{\link[igraph]{graph.adjacency}}
}
