% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMRF_adjacency_cont.R
\name{getMRF_adjacency_cont}
\alias{getMRF_adjacency_cont}
\title{Extract predicted adjacency matrices for magnitudes of a continuous covariate}
\usage{
getMRF_adjacency_cont(data, MRF_mod, covariate, plot_booted_coefs)
}
\arguments{
\item{data}{Dataframe. The input data where the
left-most variables are binary occurrences that are represented by nodes in the graph}

\item{MRF_mod}{A fitted \code{MRFcov} object}

\item{covariate}{Character representing the factor covariate name}

\item{plot_booted_coefs}{Logical. If \strong{TRUE}, mean interaction coefficients,
taken as output from a \code{bootstrap_MRF} object supplied as \code{MRF_mod},
will be plotted. Default is \strong{FALSE}}
}
\value{
A list of weighted, undirected \code{igraph} adjacency matrices (one matrix
for each observed quantile of the continuous covariate, where the observed
quantiles are the \code{min}, \code{median} and \code{max})
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}} models to
create weighted, undirected \code{igraph} adjacency matrices for different magnitudes
of a specified \code{numeric} covariate.
}
\details{
Observed values of the specified \code{covariate} are extracted by
name matching of \code{colnames(data)}. Interaction parameters from \code{MRF_mod} are
extracted and converted into predicted weighted, undirected adjacency matrices
at the covariate's \code{min}, \code{median} and \code{max} observed values
using \code{\link[igraph]{graph.adjacency}}.
}
\examples{
data("Bird.parasites")
CRFmod <- MRFcov(data = Bird.parasites, n_nodes = 4, lambda1 = 0.5)
getMRF_adjacency_cont(data = Bird.parasites, MRF_mod = CRFmod, covariate = 'scale.prop.zos')


}
\seealso{
\code{\link{MRFcov}}
}
