% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMRF_adjacency_factor.R
\name{getMRF_adjacency_factor}
\alias{getMRF_adjacency_factor}
\title{Extract predicted adjacency matrices for different factor levels}
\usage{
getMRF_adjacency_factor(MRF_mod, covariate, base_contrast_name,
  plot_booted_coefs, threshold)
}
\arguments{
\item{MRF_mod}{A fitted \code{MRFcov} object}

\item{covariate}{Character representing the factor covariate name}

\item{base_contrast_name}{Character representing the name of \code{covariate}'s base contrast
level}

\item{plot_booted_coefs}{Logical. If \strong{TRUE}, mean interaction coefficients,
taken as output from a \code{bootstrap_MRF} object supplied as \code{MRF_mod},
will be plotted. Default is \strong{FALSE}}

\item{threshold}{Numeric. The coefficient threshold used for conditional interaction prediction.
If regression coefficients for both nodes are below this threshold for a specific factor level,
both nodes are considered to be predicted as absent from that level and no interactions are predicted.
Likewise, if only one node is considered absent, then no positive interactions can occur. Default is
\code{-0.99}}
}
\value{
A list of weighted, undirected \code{igraph} adjacency matrices (one matrix for each
level of the factor covariate)
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}} models to
create weighted, undirected \code{igraph} adjacency matrices for each level
of a specified \code{factor} covariate.
}
\details{
Interaction parameters from \code{MRF_mod} are extracted and converted into
weighted, undirected adjacency matrices using \code{\link[igraph]{graph.adjacency}}.
Because base contrast levels of factors are not included in the regression design matrix,
the character name \code{base_contrast_name} must be provided so that names of the returned
list can properly match factor level names.
}
\examples{
\dontrun{
data("Dipping.survey")
CRFmod <- MRFcov(data = Dipping.survey, n_nodes = 16,
                lambda1 = 4, lambda2 = 1)
getMRF_adjacency_factor(MRF_mod = CRFmod, covariate = 'dipping.round',
                 base_contrast_name = '2')}

}
\seealso{
\code{\link{MRFcov}}
}
