% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMRF_central.R
\name{getMRF_central}
\alias{getMRF_central}
\title{Extract predicted eigenvalue centralities for a given dataset}
\usage{
getMRF_central(data, MRF_mod)
}
\arguments{
\item{data}{Dataframe. The sample data where the
left-most variables are variables that are represented by nodes in the graph}

\item{MRF_mod}{A fitted \code{MRFcov} object}
}
\value{
A \code{vector} containing predicted \code{igraph} eigenvalue centrality scores
for each node in \code{data}
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}}
and \code{\link{bootstrap_MRF}} models to
generate linear predictions for the nodes in \code{data} and
calculate centrality from a weighted adjacency matrix.
}
\details{
Interaction parameters from \code{MRF_mod} are predicted using \code{data}
and converted into a weighted, undirected adjacency matrix
using \code{\link[igraph]{graph.adjacency}}. Eigenvalue centrality scores are then
extracted using \code{\link[igraph]{eigen_centrality}}
}
\examples{
data("Bird.parasites")
CRFmod <- MRFcov(data = Bird.parasites, n_nodes = 4, family = "binomial")
getMRF_central(data = Bird.parasites[1:300, ], MRF_mod = CRFmod)


}
\seealso{
\code{\link{MRFcov}}, \code{\link[igraph]{eigen_centrality}}
}
