% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMRF_hm.R
\name{plotMRF_hm}
\alias{plotMRF_hm}
\title{Plot MRF interaction parameters as a heatmap}
\usage{
plotMRF_hm(MRF_mod, node_names, main, plot_booted_coefs)
}
\arguments{
\item{MRF_mod}{A fitted \code{MRFcov}, \code{\link[rosalia]{rosalia}} or \code{bootstrap_MRF}
object}

\item{node_names}{A character vector of species names for axis labels. Default
is to use rownames from the \code{MRFcov$graph} slot}

\item{main}{An optional character title for the plot}

\item{plot_booted_coefs}{Logical. If \code{TRUE}, interaction coefficients,
taken as output from a \code{bootstrap_MRF} object supplied as \code{MRF_mod},
will be plotted. Default is \code{FALSE}}
}
\value{
A \code{ggplot2} object
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}} models to
plot a heatmap of node interaction coefficients.
}
\details{
Interaction parameters from \code{MRF_mod} are plotted as a heatmap, where
red colours indicate positive interactions and blue indicate negative interactions
}
\examples{
\dontrun{
data("Bird.parasites")
CRFmod <- MRFcov(data = Bird.parasites,
                n_nodes = 4, lambda1 = 0.5)
plotMRF_hm(MRF_mod = CRFmod)}

}
\seealso{
\code{\link{MRFcov}}, \code{\link{bootstrap_MRF}}, \code{\link[rosalia]{rosalia}}
}
