% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMRF_hm_factor.R
\name{plotMRF_hm_factor}
\alias{plotMRF_hm_factor}
\title{Plot heatmaps of varying interaction coefficients across factor levels}
\usage{
plotMRF_hm_factor(MRF_mod, node_names, covariate, base_contrast_name, main,
  n_plot_columns, plot_booted_coefs, threshold)
}
\arguments{
\item{MRF_mod}{A fitted \code{MRFcov} object}

\item{node_names}{A character vector of species names for axis labels. Default
is to use rownames from the \code{MRFcov$graph} slot}

\item{covariate}{Character representing the factor covariate name}

\item{base_contrast_name}{Character representing the name of \code{covariate}'s base contrast
level}

\item{main}{An optional character title for the plot}

\item{n_plot_columns}{An optional integer specifying the number of columns to use
for plot facetting}

\item{plot_booted_coefs}{Logical. If \strong{TRUE}, mean interaction coefficients,
taken as output from a \code{bootstrap_MRF} object supplied as \code{MRF_mod},
will be plotted. Default is \strong{FALSE}}

\item{threshold}{Numeric. The coefficient threshold used for conditional interaction plotting.
If regression coefficients for both nodes are below this threshold for a specific factor level,
both nodes are considered to be predicted as absent from that level and no interactions will be plotted.
Likewise, if only one node is considered absent, then no positive interactions will be shown. Default is
\code{-0.99}}
}
\value{
A \code{ggplot2} object
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}} models to
plot heatmaps of node interaction coefficients across varying levels
of a specified \code{factor} covariate.
}
\details{
Interaction parameters from \code{MRF_mod} are plotted as heatmaps for each level of
\code{covariate}, with red colours indicating positive interactions and
blue indicating negative interactions. Because base contrast levels of factors are not
included in regression the design matrix, the character name \code{base_contrast_name}
must be provided
}
\examples{
\dontrun{
data("Dipping.survey")
CRFmod <- MRFcov(data = Dipping.survey, n_nodes = 16,
                lambda1 = 4, lambda2 = 1)
plotMRF_hm_factor(MRF_mod = CRFmod, covariate = 'dipping.round',
                 base_contrast_name = '2')}

}
\seealso{
\code{\link{MRFcov}}
}
