% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMRF_net_cont.R
\name{plotMRF_net_cont}
\alias{plotMRF_net_cont}
\title{Plot networks of varying interaction coefficients for a continuous covariate}
\usage{
plotMRF_net_cont(data, MRF_mod, node_names, covariate, main, plot_booted_coefs)
}
\arguments{
\item{data}{Dataframe. The input data where the
left-most variables are binary occurrences that are represented by nodes in the graph}

\item{MRF_mod}{A fitted \code{MRFcov} or \code{bootstrap_MRF} object}

\item{node_names}{A character vector of species names for axis labels. Default
is to use rownames from the \code{MRFcov$graph} slot}

\item{covariate}{Character representing the continuous covariate name}

\item{main}{An optional character title for the plot}

\item{plot_booted_coefs}{Logical. If \strong{TRUE}, mean interaction coefficients,
taken as output from a \code{bootstrap_MRF} object supplied as \code{MRF_mod},
will be plotted. Default is \strong{FALSE}}
}
\value{
A plot object
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}} models to
plot networks of node interaction coefficients across observed magnituedes of a
a specified \code{numeric} covariate.
}
\seealso{
\code{\link{MRFcov}}, \code{bootstrap_MRF}
}
