% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRFcov.R
\name{MRFcov}
\alias{MRFcov}
\title{Binary Markov Random Fields with covariates}
\usage{
MRFcov(data, lambda1, lambda2, separate_min, prep_covariates, n_nodes, n_cores,
  n_covariates)
}
\arguments{
\item{data}{Dataframe. The input data where the \code{n_nodes}
left-most variables are binary occurrences to be represented by nodes in the graph}

\item{lambda1}{Positve numeric. To be used as the l1−regularization parameter}

\item{lambda2}{Numeric (>= 0). Value for l2−regularization, where larger values lead
to stronger shrinking of coefficient magnitudes. Default is 0, but larger values
may be necessary for large or particularly sparse datasets}

\item{separate_min}{Logical. If \strong{TRUE}, interaction coefficients will use the minimum absolute value of
the corresponding parameter estimates, which are taken from separate logistic regressions,
in the symmetric postprocessed coefficient matrix. Else use the maximum
Default is \strong{FALSE}}

\item{prep_covariates}{Logical. If \strong{TRUE}, covariate columns will be cross-multiplied
with nodes to prep the dataset for MRF models. Note this is only useful when additional
covariates are provided. Therefore, if \code{n_nodes < ncol(data)},
default is \strong{TRUE}. Otherwise, default is \strong{FALSE}. See
\code{\link{prep_MRF_covariates}} for more information}

\item{n_nodes}{Positive integer. The index of the last column in \code{data}
which is represented by a node in the final graph. Columns with index
greater than n_nodes are taken as covariates. Default is the number of
columns in \code{data}, corresponding to no additional covariates}

\item{n_cores}{Positive integer. The number of cores to spread the job across using
\code{\link[parallel]{makePSOCKcluster}}. Default is 1 (no parallelisation)}

\item{n_covariates}{Positive integer. The number of covariates in \code{data}, before cross-multiplication}
}
\value{
A \code{list} of six objects:
\itemize{
   \item \code{graph}: Estimated parameter matrix of interaction effects
   \item \code{intercepts}: Estimated parameter vector of node intercepts
   \item \code{results}: \code{list} of length \code{n_nodes} containing results of
   individual regularized logistic regressions
   \item \code{indirect_coefs}: \code{list} containing matrices of indirect effects of
   each covariate on node interactions
   \item \code{direct_coefs}: \code{matrix} of direct covariate effects on
   node occurrence probabilities
   \item \code{param_names}: Character string of covariate parameter names
   }
}
\description{
This function is the workhorse of the \code{MRFcov} package, running
separate penalized logistic regressions for each node to estimate parameters of binary
Markov Random Fields (MRF) graphs. Covariates can be included
(a class of models known as Conditional Random Fields; CRF), to estimate
how interactions vary across covariate magnitudes.
}
\details{
Separate penalized logistic regressions are used to approximate
MRF parameters, where the regression for species \code{j} includes an
intercept and beta coefficients for the presence-absence of all other
species (\code{/j}) in \code{data}. If covariates are included, beta coefficients
are also estimated for the effect of the covariate on \code{j} and the
effects of the covariate on interactions between \code{j} and all other species
(\code{/j}). Because the number of parameters quickly increases with increasing
numbers of species and covariates, LASSO penalization is used to regularize
regressions based on values of regularization parameters \code{lambda1} and
\code{lambda2}. See \code{\link[penalized]{penalized}} for details of
the penalization process
}
\examples{
\dontrun{
data("Bird.parasites")
CRFmod <- MRFcov(data = Bird.parasites, n_nodes = 4,
          lambda1 = 0.5)}

}
\references{
Ising, E. (1925). Beitrag zur Theorie des Ferromagnetismus.
Zeitschrift für Physik A Hadrons and Nuclei, 31, 253-258.\cr\cr
Cheng, J., Levina, E., Wang, P. & Zhu, J. (2014).
A sparse Ising model with covariates. (2012). Biometrics, 70, 943-953.\cr\cr
Sutton C, McCallum A. An introduction to conditional random fields.
Foundations and Trends in Machine Learning 4, 267-373.
}
\seealso{
\href{https://www.doria.fi/bitstream/handle/10024/124199/ThesisOscarLindberg.pdf?sequence=2}{Lindberg (2016)}
and \href{http://homepages.inf.ed.ac.uk/csutton/publications/crftut-fnt.pdf}{Sutton & McCallum (2012)}
for overviews of Conditional Random Fields, \code{\link[penalized]{penalized}} for
details of penalized logistic regressions
}
