% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMRF_hm.R
\name{plotMRF_hm}
\alias{plotMRF_hm}
\title{Plot MRF interaction parameters as a heatmap}
\usage{
plotMRF_hm(MRF_mod, node_names, main, plot_booted_coefs)
}
\arguments{
\item{MRF_mod}{A fitted \code{MRFcov}, \code{\link[rosalia]{rosalia}} or \code{BootstrapCoefs} object}

\item{node_names}{A character vector of species names for axis labels. Default
is to use rownames from the \code{MRFcov$graph} slot}

\item{main}{An optional character title for the plot}

\item{plot_booted_coefs}{Logical. If \strong{TRUE}, interaction coefficients,
taken as output from a \code{bootstrap_MRF} object supplied as \code{MRF_mod},
will be plotted. Default is \strong{FALSE}}
}
\value{
A \code{ggplot2} object
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}} models to
plot a heatmap of node interaction coefficients.
}
\seealso{
\code{\link{MRFcov}}, \code{\link{bootstrap_MRF}}
}
