% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMRF_hm_cont.R
\name{plotMRF_hm_cont}
\alias{plotMRF_hm_cont}
\title{Plot heatmaps of varying interaction coefficients for a continuous covariate}
\usage{
plotMRF_hm_cont(data, MRF_mod, node_names, covariate, main, n_plot_columns)
}
\arguments{
\item{data}{Dataframe. The input data where the
left-most variables are binary occurrences that are represented by nodes in the graph}

\item{MRF_mod}{A fitted \code{MRFcov} or \code{bootstrap_MRF} object}

\item{node_names}{A character vector of species names for axis labels. Default
is to use rownames from the \code{MRFcov$graph} slot}

\item{covariate}{Character representing the continuous covariate name}

\item{main}{An optional character title for the plot}

\item{n_plot_columns}{An optional integer specifying the number of columns to use
for plot facetting}
}
\value{
A \code{ggplot2} object
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}} models to
plot heatmaps of node interaction coefficients across observed magnituedes of a
a specified \code{numeric} covariate.
}
\details{
Observed values of the specified \code{covariate} are extracted by
name matching of \code{colnames(data)}. Interaction parameters from \code{MRF_mod} are
are then predicted at \code{quantile(probs = c(0, 0.25, 0.75, 1))} of observed values, where
red colours indicate positive interactions and blue indicate negative interactions
}
\seealso{
\code{\link{MRFcov}}, \code{bootstrap_MRF}
}
