% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMRF_net_cont.R
\name{plotMRF_net_cont}
\alias{plotMRF_net_cont}
\title{Plot networks of varying interaction coefficients for a continuous covariate}
\usage{
plotMRF_net_cont(data, MRF_mod, node_names, covariate, main, cutoff, plot)
}
\arguments{
\item{data}{Dataframe. The input data where the
left-most variables are binary occurrences that are represented by nodes in the graph}

\item{MRF_mod}{A fitted \code{MRFcov} or \code{bootstrap_MRF} object}

\item{node_names}{A character vector of species names for axis labels. Default
is to use rownames from the \code{MRFcov$graph} slot}

\item{covariate}{Character representing the continuous covariate name}

\item{main}{An optional character title for the plot}

\item{cutoff}{Positive numeric. Interaction coefficients whose absolute values are
below this threshold will not be plotted. This is useful when many weak interactions
tend to create cluttered and uninterpetable network plots. Default is \code{0}}

\item{plot}{Logical. If \code{TRUE}, returns a gridded plot object showing predicted
networks across three quantiles of the covariate (minimum, median, and maximum).
If \code{FALSE}, returns three weighted, undirected adjacency matrices that can be plotted
using functions in \code{\link[igraph]{igraph}}. \code{TRUE} is default}
}
\value{
Either a plot object (if \code{plot = TRUE}) or a list of three weighted,
undirected \code{igraph} adjacency matrices
}
\description{
This function uses outputs from fitted \code{\link{MRFcov}} models to
plot networks of node interaction coefficients across observed magnitudes of a
a specified continuous covariate.
}
\details{
Observed values of the specified \code{covariate} are extracted by
name matching of \code{colnames(data)}. Interaction parameters from \code{MRF_mod} are
are then predicted at \code{min}, \code{median} and \code{max} of observed values, where
red edge colours indicate positive interactions, blue indicate negative interactions, and
the widths of edges indicate strengths of interactions
}
\examples{
data("Bird.parasites")
CRFmod <- MRFcov(data = Bird.parasites, n_nodes = 4, lambda1 = 0.5)
plotMRF_net_cont(data = Bird.parasites, MRF_mod = CRFmod, covariate = 'scale.prop.zos')

}
\seealso{
\code{\link{MRFcov}}, \code{\link{bootstrap_MRF}}
}
