% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MRF.R
\name{predict_MRF}
\alias{predict_MRF}
\title{Predict training observations from fitted MRFcov models}
\usage{
predict_MRF(data, MRF_mod, prep_covariates = TRUE)
}
\arguments{
\item{data}{Dataframe. The input data to be predicted, where the \code{n_nodes}
left-most variables are binary occurrences to be represented by nodes in the graph.
Colnames from this sample dataset must exactly match the colnames in the dataset that
was used to fit the \code{MRF_mod}}

\item{MRF_mod}{A fitted \code{\link{MRFcov}} model object}

\item{prep_covariates}{Logical flag stating whether to prep the dataset
by cross-multiplication (\code{TRUE} by default; \code{FALSE} when used in other functions)}
}
\value{
A \code{matrix} containing predictions for each observation in \code{data}:
}
\description{
This function calculates linear predictors for node observations
using equations from a \code{\link{MRFcov}}.
}
\details{
Observations for nodes in \code{data} are predicted using linear predictions
from \code{MRF_mod}. If \code{family = "binomial"}, a second element containing binary
predictions for nodes is returned.
}
\examples{
\dontrun{
data("Bird.parasites")
# Fit a model to a subset of the data (training set)
CRFmod <- MRFcov(data = Bird.parasites[1:300, ], n_nodes = 4, family = "binomial")

# If covariates are included, prep the dataset for gathering predictions
prepped_pred <- prep_MRF_covariates(Bird.parasites[301:nrow(Bird.parasites), ], n_nodes = 4)

# Predict occurrences for the remaining subset (test set)
predictions <- predict_MRF(data = prepped_pred, MRF_mod = CRFmod)

# Visualise predicted occurrences for nodes in the test set
predictions$Binary_predictions}

}
\seealso{
\code{\link{MRFcov}}, \code{\link{cv_MRF}}
}
