% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MRF.R
\name{predict_MRF}
\alias{predict_MRF}
\title{Predict training observations from fitted MRFcov models}
\usage{
predict_MRF(data, MRF_mod, n_nodes)
}
\arguments{
\item{data}{Dataframe. The input data to be predicted, where the \code{n_nodes}
left-most variables are binary occurrences to be represented by nodes in the graph}

\item{MRF_mod}{A fitted \code{\link{MRFcov}} model object}

\item{n_nodes}{Positive integer. The index of the last column in data
which is represented by a node in the final graph. Columns with index
greater than \code{n_nodes} are taken as covariates. Default is the number of
columns in data, corresponding to no additional covariates}
}
\value{
A \code{list} of two objects:
\itemize{
   \item \code{binary_predictions}: Data.frame containing the binary predicted
   values for node variable observations in data
   \item \code{linear_predictions}: Data.frame containing the linear predicted values
   for node variable observations in data
   }
}
\description{
This function calculates linear predictors for node observations
using equations from a \code{\link{MRFcov}}.
}
\details{
Binary occurrences for nodes in \code{data} are predicted using linear predictions
from \code{MRF_mod}. This function is not designed to be called directly, but is used by
other functions in the package (namely \code{\link{MRFcov}}, \code{\link{cv_MRF}}, and
\code{\link{bootstrap_MRF}})
}
\seealso{
\code{\link{MRFcov}}, \code{\link{cv_MRF}}
}
